backend/requirements.txt
Flask==2.3.0
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.4
Flask-JWT-Extended==4.5.2
Flask-CORS==4.0.0
Flask-RESTful==0.3.10
Flask-Marshmallow==0.15.0
marshmallow-sqlalchemy==0.29.0
bcrypt==4.0.1
python-dotenv==1.0.0
gunicorn==21.2.0
pytest==7.4.0
pytest-cov==4.1.0
